fix: drop tables before start refactor: refactor pet tests
