+++ src/main/resources/application.properties
@@ -10,13 +10,13 @@ spring.datasource.username=${POSTGRES_USER}
 spring.datasource.password=${POSTGRES_PASSWORD}
 
 DB_NAME=pethotel
-POSTGRES_USER=admin
-POSTGRES_PASSWORD=admin_pwd
+POSTGRES_USER=pet_admin
+POSTGRES_PASSWORD=pet_admin_pwd
 DB_HOST=localhost
 DB_PORT=5432
 #---
-#spring.config.activate.on-profile=ci,test
-#spring.datasource.driverClassName=org.h2.Driver
-#spring.datasource.url=jdbc:h2:mem:pethotel
-#spring.datasource.username=test
-#spring.datasource.password=test
+spring.config.activate.on-profile=ci,test
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.url=jdbc:h2:mem:pethotel
+spring.datasource.username=test
+spring.datasource.password=test
